re-organize the Atom automated test structure to match the new parallel + batched test structures
