+++ AutomatedTesting/Gem/PythonTests/Atom/CMakeLists.txt
