+++ AutomatedTesting/Gem/PythonTests/Atom/__init__.py
