+++ AutomatedTesting/Gem/PythonTests/Atom/atom_utils/atom_component_helper.py
