+++ AutomatedTesting/Gem/PythonTests/Atom/atom_utils/atom_constants.py
