+++ AutomatedTesting/Gem/PythonTests/Atom/atom_utils/screenshot_utils.py
