+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/AreaLight_1.ppm
