+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/AreaLight_2.ppm
