+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/AreaLight_3.ppm
