+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/AreaLight_4.ppm
