+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/AreaLight_5.ppm
