+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/SpotLight_1.ppm
