+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/SpotLight_2.ppm
