+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/SpotLight_3.ppm
