+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/SpotLight_4.ppm
