+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/SpotLight_5.ppm
