+++ AutomatedTesting/Gem/PythonTests/Atom/golden_images/SpotLight_6.ppm
