+++ AutomatedTesting/Gem/PythonTests/Atom/tests/TestAssets/ShaderAssetBuilder/DependencyValidation.shader.txt
@@ -2,7 +2,7 @@
 {
     "Source" : "DependencyValidation.azsl",
 
-    "DepthStencilState" : { 
+    "DepthStencilState" : {
         "Depth" : { "Enable" : false, "CompareFunc" : "GreaterEqual" }
     },
 
@@ -22,5 +22,5 @@
         }
       ]
     }
-    
+
 }
