+++ AutomatedTesting/Gem/PythonTests/Atom/tests/TestAssets/ShaderAssetBuilder/Test3Color.azsli.txt
@@ -12,5 +12,5 @@
 
 float4 GetTest3Color(float4 color)
 {
-    return color * 0.13; 
+    return color * 0.13;
 }
